requirements.txt
websocket-client
Pillow
flask
